requirements.txt
fastapi==0.88.0
numpy==1.21.6
uvicorn==0.20.0
image==1.5.33
tensorflow-cpu==2.9.2
tensorflow-hub==0.12.0